# Refactor RealmLog