--- a/realm/realm-library/src/androidTest/java/io/realm/NotificationsTest.java
+++ b/realm/realm-library/src/androidTest/java/io/realm/NotificationsTest.java
@@ -47,6 +47,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 import io.realm.entities.AllTypes;
 import io.realm.entities.Dog;
+import io.realm.log.LogLevel;
 import io.realm.log.Logger;
 import io.realm.log.RealmLog;
 import io.realm.rule.RunInLooperThread;
@@ -1180,9 +1181,11 @@ public class NotificationsTest {
         final AtomicBoolean warningLogged = new AtomicBoolean(false);
         final TestHelper.TestLogger testLogger = new TestHelper.TestLogger() {
             @Override
-            public void warn(Throwable t, String message, Object... args) {
+            public void log(int level, String tag, Throwable throwable, String message) {
                 assertTrue(message.contains("Mixing asynchronous queries with local writes should be avoided."));
-                warningLogged.set(true);
+                if (level == LogLevel.WARN) {
+                    warningLogged.set(true);
+                }
             }
         };
         RealmLog.add(testLogger);