--- a/realm/realm-library/src/androidTest/java/io/realm/RealmLogTests.java
+++ b/realm/realm-library/src/androidTest/java/io/realm/RealmLogTests.java
@@ -112,7 +112,6 @@ public class RealmLogTests {
 
         @Override
         public void info(Throwable throwable, String message, Object... args) {
-
         }
 
         @Override