--- a/realm/realm-library/src/main/cpp/io_realm_internal_Util.cpp
+++ b/realm/realm-library/src/main/cpp/io_realm_internal_Util.cpp
@@ -19,7 +19,6 @@
 #include <realm/string_data.hpp>
 #include <realm/unicode.hpp>
 
-#include "io_realm_log_LogLevel.h"
 #include "mem_usage.hpp"
 #include "util.hpp"
 
@@ -29,16 +28,6 @@ using std::string;
 #if defined(_MSC_VER) && defined(_DEBUG) && defined(USE_VLD)
     #include "C:\\Program Files (x86)\\Visual Leak Detector\\include\\vld.h"
 #endif
-
-// used by logging
-int trace_level = 0;
-jclass realmlog_class;
-jmethodID log_trace;
-jmethodID log_debug;
-jmethodID log_info;
-jmethodID log_warn;
-jmethodID log_error;
-jmethodID log_fatal;
 
 const string TABLE_PREFIX("class_");
 
@@ -58,13 +47,6 @@ JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM* vm, void*)
         java_lang_float_init  = env->GetMethodID(java_lang_float, "<init>", "(F)V");
         java_lang_double      = GetClass(env, "java/lang/Double");
         java_lang_double_init = env->GetMethodID(java_lang_double, "<init>", "(D)V");
-        realmlog_class        = GetClass(env, "io/realm/log/RealmLog");
-        log_trace             = env->GetStaticMethodID(realmlog_class, "trace", "(Ljava/lang/String;[Ljava/lang/Object;)V");
-        log_debug             = env->GetStaticMethodID(realmlog_class, "debug", "(Ljava/lang/String;[Ljava/lang/Object;)V");
-        log_info              = env->GetStaticMethodID(realmlog_class, "info", "(Ljava/lang/String;[Ljava/lang/Object;)V");
-        log_warn              = env->GetStaticMethodID(realmlog_class, "warn", "(Ljava/lang/String;[Ljava/lang/Object;)V");
-        log_error             = env->GetStaticMethodID(realmlog_class, "error", "(Ljava/lang/String;[Ljava/lang/Object;)V");
-        log_fatal             = env->GetStaticMethodID(realmlog_class, "fatal", "(Ljava/lang/String;[Ljava/lang/Object;)V");
     }
 
     return JNI_VERSION_1_6;
@@ -83,21 +65,6 @@ JNIEXPORT void JNI_OnUnload(JavaVM* vm, void*)
     }
 }
 
-JNIEXPORT void JNICALL Java_io_realm_internal_Util_nativeSetDebugLevel(JNIEnv*, jclass, jint level)
-{
-    /**
-     * level should match one of the levels defined in LogLevel.java
-     * ALL = 1
-     * TRACE = 2
-     * DEBUG = 3
-     * INFO = 4
-     * WARN = 5
-     * ERROR = 6
-     * FATAL = 7
-     * OFF = 8
-     */
-    trace_level = level;
-}
 
 JNIEXPORT jlong JNICALL Java_io_realm_internal_Util_nativeGetMemUsage(JNIEnv*, jclass)
 {