… CHANGELOG.md
… examples/objectServerExample/src/main/java/io/realm/examples/objectserver/MyApplication.java
… realm/realm-library/src/main/cpp/CMakeLists.txt
… realm/realm-library/src/main/cpp/jni_util/log.cpp
… realm/realm-library/src/main/cpp/jni_util/log.hpp
… realm/realm-library/src/main/cpp/util.cpp
… realm/realm-library/src/main/cpp/util.hpp
… realm/realm-library/src/main/java/io/realm/Realm.java
… realm/realm-library/src/main/java/io/realm/internal/Util.java
… realm/realm-library/src/main/java/io/realm/log/AndroidLogger.java
… realm/realm-library/src/main/java/io/realm/log/Logger.java
… realm/realm-library/src/main/java/io/realm/log/RealmLog.java
… realm/realm-library/src/main/java/io/realm/log/RealmLogger.java
… realm/realm-library/src/objectServer/java/io/realm/SyncManager.java