The motivation:
* The previous implementaion for JNI log needs calls like:
  JNI -> Java -> JNI. It is not quite effecient for the most common case
  -- log in JNI and sync.
* The previous log levels are bit confusing.
* util.cpp gets messy.
* The Logger class in Java gets confusing with the Logger from
  core/sync.

So this commit makes below changes:
* Adds cpp namespace jni_impl and jni_util, and make log relevant code
  belong to these two namepsaces.
* Moves the RealmLog to JNI.
* Implenment a default Android logger in native code.
* Add tag support for realm-java Logger. Java side is using tag
  "REALM_JAVA", JNI side is using tag "REALM_JNI" and sync is using tag
  "REALM_CORE".
* All tags share the same log level.
* Some cleanups.

Benchmark:

|            |  Before NokiaX2 - 4.3           | After NokiaX2 - 4.3  |
| ------------- |:-------------:| -----:|
| readDouble      | 2.37 μs | 2.39 μs |
| readLong     | 2.53 μs      |  2.54 μs |
| readString | 10.36 μs      |   9.51 μs |