@beeender 2016-10-17
@realm/java This is ready for review. I am checking https://github.com/realm/realm-java/issues/3528 , will update this PR if there is anything needed for that.