@beeender 2016-10-21
@realm/java help to review again.