@beeender 2016-10-24
@cmelchior For `GetEnv` vs `AttachCurrentThread` on a thread which has a vm attached:
On Nokia X2 (davik) 470ns vs 576 ns
On Emulator 6.0 (art) 36ns vs 52ns