@beeender 2016-10-24
@realm/java Please review again